Use `as` instead of `url` to determine URL newness
Currently, using `as` will cause the router to think the URL is not
changing in the case where you're re-rendering the same page with a
different route. This would of course be an issue for custom servers
only.

This should be an invisible change for non-custom-server users, since
`as` is defaulted to `url` if not set.

This should resolve #3065.